Resolved Dan Marsden's issues:
* Separated the class edit into a single class name feed_edit -> fixes #6
* Added namespaces for editfeed and form/feed_edit -> fixes #5
* Added thirdpartylibs.xml -> fixes #3
* Added a check at the start of block_rss_thumbnails::get_content so the method will no longer compute twice -> fixes #4
