+++ version.php
@@ -24,7 +24,8 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->version    = 2022102501;        // The current plugin version (Date: YYYYMMDDXX).
+$plugin->version    = 2022111004;        // The current plugin version (Date: YYYYMMDDXX).
 $plugin->requires   = 2012112900;        // Requires this Moodle version.
+$plugin->release    = 1.0;
 $plugin->component  = 'block_rss_thumbnails'; // Full name of the plugin (used for diagnostics).
 $plugin->dependencies = [];
